requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
backports.zoneinfo==0.2.1
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
